+++ spec/integration/generate_all_endpoints_spec.rb
@@ -19,6 +19,8 @@ RSpec.describe "Generate an OpenAPI document for all endpoints", :rails_app do
 
     expect(document["paths"].keys).to contain_exactly(
       "/api/activity_logs",
+      "/api/binding_helpers/chain", "/api/binding_helpers/create",
+      "/api/binding_helpers/kwargs",
       "/api/case_branches/show",
       "/api/constant_references/execute", "/api/constant_references/missing",
       "/api/constant_references/non_compatible", "/api/constant_references/pattern",